requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.0.0